Separates the Reconciliation from utils
